Use '<profile>.end' if it exists
If a profile has a corresponding 'end' profile, its contents are applied
in reverse order after the topological sort.

e.g. if the profiles are: 'a', 'b', 'c', and there also exist 'b.end'
and 'c.end', then the following profiles will be applied in the
following order: 'a', 'b', 'c', 'c.end', 'b.end'.