Modify profile layout
Each profile is now contained in a single directory, with optional
'parents', 'up' and 'down' entries that contain the parent information
as well as parts of the profiles to apply in reverse topologically
sorted order and topologically sorted order respectively. This reflects
the intuition that the 'up' profiles are applied going bottom up and the
'down' profiles are applied going top down.

This reduces clutter in profile collections and as a side effect
simplifies the implementation of 'stratum graph'.